backend/django_auth/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# CORS (for React frontend)
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Database
psycopg2-binary==2.9.10  # PostgreSQL (production)

# Email backend (development)
django-extensions==3.2.3

# API Documentation
drf-spectacular==0.26.5

# Development tools
black==23.9.1
flake8==6.1.0

# Testing
pytest-django==4.5.2
factory-boy==3.3.0

setuptools
Pillow
dj-database-url
django-redis>=5.4.0
